Normalize the path before "list"
No matter how the path is formatted in the database like "a/b" or "a\b",
the listed result will always use the format of the server's runtime
environment.

This will allow Windows to correctly read the database created by Linux
or the opposite.

An alternative way is to store the same format no matter which runtime
environment the server is under. It seems the first way is eaiser to be
implemented.